Fix the regular expression with decimal number for calc-views and urls

+++ zqxt_views/calc/views.py
@@ -9,22 +9,22 @@ from django.http import HttpResponse
 def add(Request):
     a = Request.GET['a']
     b = Request.GET['b']
-    c = int(a) + int(b)
+    c = float(a) + float(b)
     return HttpResponse(str(c))
     #URL format is:http://127.0.0.1:8000/add/?a=111&b=333
 
 def add2(Request,a,b):
-    c = int(a) + int(b)
+    c = float(a) + float(b)
     return HttpResponse(str(c))
     #URL format is:http://127.0.0.1:8000/add/111/333
 
 def subtract(Request,a,b):
-    w = int(a) - int(b)
+    w = float(a) - float(b)
     return HttpResponse((w))
     #URL format is:http://127.0.0.1:8000/subtract/111/333
 
 def multiply(Request,a,b):
-    h = int(a) * int(b)
+    h = float(a) * float(b)
     return HttpResponse(str(h))
     #URL format is:http://127.0.0.1:8000/multiply/111/333
 
@@ -32,7 +32,7 @@ def divide(request,a,b):
     if float(b) == 0:
         return HttpResponse(u'0不能作为除数')
     else:
-        k = int(a) / int(b)
+        k = float(a) / float(b)
     return HttpResponse(str(k))
     #URL format is:http://127.0.0.1:8000/divide/111/333
